+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Fixed a bug in `TLapJoint` and `XLapJoint` where the `cut_plane_bias` parameter was not passed to the `_create_negative_volumes()` method after its signature was changed.
 * Replaced `JackRafterCut` and `Lap` with their Proxy counterparts in `LLapJoint` and `TLapJoint`.
+* Renamed `dovetail_volume_from_params_and_beam` to `volume_from_params_and_beam` in `DovetailTenon` and `DovetailMortise`.
+* Renamed `from_plane_and_beam` method in `compas_timber.fabrication.Tenon` to `from_frame_and_beam`.
+* Updated the alternative constructor of`compas_timber.fabrication.Tenon` so that the rotation parameter is calculated automatically.
+* Updated `TDovetailJoint` and `TenonMortiseJoint` from `compas_timber.connections` so that they use `House` and `HouseMortise`if needed.
 
 ### Removed